+++ src/drivers/opengles/opengl_shader/lv_opengl_shader_internal.h
@@ -6,11 +6,11 @@
 #ifndef LV_OPENGL_SHADER_INTERNAL_H
 #define LV_OPENGL_SHADER_INTERNAL_H
 
-#include "../../../lv_conf_internal.h"
+#include "../../../../lv_conf_internal.h"
 
 #if LV_USE_GLTF
-#include "../../../drivers/opengles/lv_opengles_private.h"
-#include "../../../drivers/opengles/lv_opengles_debug.h"
+#include "../lv_opengles_private.h"
+#include "../lv_opengles_debug.h"
 #include "../../../misc/lv_types.h"
 #include "../../../misc/lv_rb_private.h"
