+++ src/drivers/opengles/opengl_shader/lv_opengl_shader_program.c
@@ -11,8 +11,8 @@
 
 #if LV_USE_GLTF
 
-#include "../../../drivers/opengles/lv_opengles_private.h"
-#include "../../../drivers/opengles/lv_opengles_debug.h"
+#include "../lv_opengles_private.h"
+#include "../lv_opengles_debug.h"
 #include "../../../misc/lv_assert.h"
 #include "../../../stdlib/lv_mem.h"
 /*********************
