+++ src/libs/gltf/gltf_view/assets/lv_gltf_view_shader.h
@@ -13,8 +13,7 @@ extern "C" {
 /*********************
  *      INCLUDES
  *********************/
-
-#include "../../opengl_shader/lv_opengl_shader_internal.h"
+#include "../../../../drivers/opengles/opengl_shader/lv_opengl_shader_internal.h"
 
 #if LV_USE_GLTF
