+++ src/libs/gltf/gltf_view/ibl/lv_gltf_ibl_sampler.c
@@ -17,7 +17,7 @@
 #include "../../../../drivers/opengles/lv_opengles_private.h"
 #include "../../../../drivers/opengles/lv_opengles_debug.h"
 
-#include "../../opengl_shader/lv_opengl_shader_internal.h"
+#include "../../../../drivers/opengles/opengl_shader/lv_opengl_shader_internal.h"
 #include "../lv_gltf_view_internal.h"
 #include "../assets/lv_gltf_view_shader.h"
