+++ src/libs/gltf/gltf_view/lv_gltf_view_shader.cpp
@@ -12,7 +12,7 @@
 #include "fastgltf/types.hpp"
 #include "../gltf_data/lv_gltf_data_internal.hpp"
 #include "../gltf_data/lv_gltf_data_internal.h"
-#include "../opengl_shader/lv_opengl_shader_internal.h"
+#include "../../../drivers/opengles/opengl_shader/lv_opengl_shader_internal.h"
 #include "../../../misc/lv_array.h"
 #include "../../../misc/lv_assert.h"
 #include "../../../misc/lv_types.h"
